Import toaster 0.1
This is a tiny shell wrapper developer by DEBAMAX for a customer, making
it possible to tweak amd64 and arm64 system images: they are duplicated
first, their partitions are mounted, and the contents can be modified at
will.

This makes it possible to move fast, as the original images don't need
to be built again to test modifications. For the PiRogue Tool Suite
project, that means we can focus on modifying Debian-provided images
instead of building them from scratch.

At the moment, support for partition tables is pretty basic, the second
partition is the root filesystem, the first partition might be mounted
on /boot/firmware (arm64 images for Raspberry Pi) or on /boot/efi (amd64
images for 64-bit PCs).

If we end up using this tool for Cloud images, at least that part will
require some modifications.