Add script to build images and a list of files to publish.
Given a RASPBERRYPI_IMG_URL, this automates downloading the image and
its checksum, validating it, then modifying it to generate two images:
 - PiRogue-OS-12-Pi3_and_Pi4
 - PiRogue-OS-12-Pi5-Experimental

Those are then compressed, checksums are generated, and the resulting
4 files are listed in a MANIFEST.txt file at the top-level, making it
trivial to publish the relevant files.

At least for the time being, no clean-up is performed. That means the
original image can be used to run ./toaster again if a new development
iteration is required.