Avoid 3 known debconf prompts during deployment.
Even if those prompts and the associated answers are already documented
in “Beginner guide - How to setup a PiRogue”, it's a bit annoying to
have to go through them every time a new PiRogue is set up.

Let's pick the right answers (moving away from the default every time),
and mark those questions as seen (i.e. asked and answered):

 - Save current IPv4 rules?                           Yes → No
 - Save current IPv6 rules?                           Yes → No
 - Should non-superusers be able to capture packets?  No → Yes

Link: https://github.com/PiRogueToolSuite/pirogue-images/issues/3